LoudnessBase in lib-builtin-effects
